requirements.txt
transformers==4.12.5
pytorch-lightning==1.5.4
pytorch-nlp==0.5.0
scikit-learn==0.24.2